INS-568 split mocks to different files. refactor handlers
